Separated atomic emission spectra into atomic.py
